+++ edsl/exceptions/agents.py
@@ -25,8 +25,10 @@ class AgentLacksLLMError(AgentErrors):
 class AgentRespondedWithBadJSONError(AgentErrors):
     pass
 
+
 class AgentNameError(AgentErrors):
     pass
+
 
 class FailedTaskException(Exception):
     def __init__(self, message, agent_response_dict):