+++ edsl/jobs/Interview.py
@@ -2,8 +2,10 @@ from __future__ import annotations
 import asyncio
 import logging
 import textwrap
-from collections import UserDict
+from collections import UserList
 from typing import Any, Type, List, Generator, Callable, List
+from collections import defaultdict
+
 from edsl import CONFIG
 from edsl.agents import Agent
 from edsl.exceptions import InterviewErrorPriorTaskCanceled, InterviewTimeoutError
@@ -14,8 +16,19 @@ from edsl.surveys import Survey
 from edsl.utilities.decorators import sync_wrapper
 from edsl.data_transfer_models import AgentResponseDict
 from edsl.jobs.Answers import Answers
-from collections import UserList
 
+from edsl.surveys.base import EndOfSurvey
+
+from edsl.jobs.buckets import ModelBuckets
+from edsl.jobs.token_tracking import TokenUsage, InterviewTokenUsage
+
+from edsl.jobs.task_management import (
+    InterviewStatusDictionary,
+    QuestionTaskCreator,
+    TasksList,
+)
+
+import traceback
 
 # create logger
 logger = logging.getLogger(__name__)
@@ -35,58 +48,7 @@ logger.addHandler(fh)
 # start loggin'
 logger.info("Interview.py loaded")
 
-
 TIMEOUT = float(CONFIG.get("API_CALL_TIMEOUT_SEC"))
-
-# class FailedTask(UserDict):
-#     def __init__(self, e: Exception = None):
-#         data = {
-#             "answer": "Failure",
-#             "comment": "Failure",
-#             "prompts": {"user_prompt": "", "sytem_prompt": ""},
-#             "exception": e,
-#         }
-#         super().__init__(data)
-
-
-class QuestionTaskCreator(UserList):
-    """Class to create and manage question tasks with dependencies."""
-
-    def __init__(self, func: Callable):
-        # self.func = func  # the function to be called to actually run the task
-        super().__init__([])
-        self.func = func
-
-    async def _run_task_async(self, question, debug) -> asyncio.Task:
-        """Runs the task asynchronously, awaiting the tasks that must be completed before this one can be run."""
-        logger.info(f"Running task for {question.question_name}")
-        try:
-            # This is waiting for the tasks that must be completed before this one can be run.
-            # This does *not* use the return_exceptions = True flag, so if any of the tasks fail,
-            # it throws the exception immediately, which is what we want.
-            await asyncio.gather(*self)
-        except asyncio.CancelledError:
-            logger.error(f"Task for {question.question_name} was cancelled, most likely because it was skipped")
-            #raise
-        except Exception as e:
-            logger.error(f"Required tasks for {question.question_name} failed: {e}")
-            # turns the parent exception into a custom exception
-            # So the task gets canceled but this InterviewErrorPriorTaskCanceled exception
-            # So we never get the question details we need.
-            raise InterviewErrorPriorTaskCanceled(
-                f"Required tasks failed for {question.question_name}"
-            ) from e
-        else:
-            logger.info(f"Tasks for {question.question_name} completed")
-            # This is the actual task that we want to run.
-            results = await self.func(question, debug)
-            return results
-
-    def __call__(self, question, debug):
-        """Creates a task that depends on the passed-in dependencies."""
-        task = asyncio.create_task(self._run_task_async(question, debug))
-        task.edsl_name = question.question_name
-        return task
 
 
 class Interview:
@@ -111,109 +73,124 @@ class Interview:
         self.verbose = verbose
         self.answers: dict[str, str] = Answers()  # will get filled in
 
-        self.dag = self.survey.dag(
-            textify=True
-        )  # the DAG tells us what questions depend on what other questions
+        self.dag = self.survey.dag(textify=True)
         self.to_index = {
             name: index for index, name in enumerate(self.survey.question_names)
         }
 
         logger.info(f"Interview instantiated")
-    
+        self.task_creators = {}
+
+    @property
+    def token_usage(self) -> InterviewTokenUsage:
+        "Determins how many tokens were used for the interview."
+        cached_tokens = TokenUsage(from_cache=True)
+        new_tokens = TokenUsage(from_cache=False)
+        for task_creator in self.task_creators.values():
+            token_usage = task_creator.token_usage()
+            cached_tokens += token_usage["cached_tokens"]
+            new_tokens += token_usage["new_tokens"]
+        return InterviewTokenUsage(
+            new_token_usage=new_tokens, cached_token_usage=cached_tokens
+        )
+
+    @property
+    def interview_status(self) -> InterviewStatusDictionary:
+        """Returns a dictionary mapping task status codes to counts"""
+        status_dict = InterviewStatusDictionary()
+        for task_creator in self.task_creators.values():
+            status_dict[task_creator.task_status] += 1
+            status_dict["number_from_cache"] += task_creator.from_cache
+        return status_dict
 
     async def async_conduct_interview(
-        self, debug: bool = False, replace_missing: bool = True
+        self,
+        model_buckets: ModelBuckets,
+        debug: bool = False,
+        replace_missing: bool = True,
     ) -> tuple["Answers", List[dict[str, Any]]]:
         """
-        Conducts an 'interview' asynchronously. An interview is:
-        - one agent
-        - one survey (so multiple questions)
-        - one model
-        - one scenario
-
-        Args:
-            debug (bool): Enable debugging mode.
-            replace_missing (bool): Replace missing answers with None.
-            threaded (bool): Flag to use threading if required.
-
-        Returns:
-            Tuple[Answers, List[Dict[str, Any]]]: The answers and a list of valid results.
+        Conducts an 'interview' asynchronously.
         """
 
-        # Each question is 1 task; this creates a list of tasks
-        # the same length as the number of questions in the survey
-        
-        tasks, invigilators = self._build_question_tasks(debug)
-        self.tasks = tasks
-        self.invigilators = invigilators
-        # when return_exceptions=False, it will just raise the exception
-        # and break the loop; otherwise it returns.
-        
-        # TODO: When debugging, we want to see the exception when they happen
-        # rather than just let them accumulate.
-        debug = False
-        if debug:
-            # If in debug mode, we want to see the exception when they happen
-            await asyncio.gather(*tasks, return_exceptions=False)
-        else:
-            await asyncio.gather(*tasks, return_exceptions=True)
-            
+        # we create both tasks and invigilators lists.
+        # this is because it's easier to extract info
+        # we need from the invigilators list when a task fails.
+        # it's challenging to get info from failed asyncio tasks.
+        self.tasks, self.invigilators = self._build_question_tasks(
+            debug=debug, model_buckets=model_buckets
+        )
+
+        await asyncio.gather(*self.tasks, return_exceptions=not debug)
 
         if replace_missing:
             self.answers.replace_missing_answers_with_none(self.survey)
 
-        valid_results = list(self._extract_valid_results(tasks, invigilators))
+        valid_results = list(self._extract_valid_results(self.tasks, self.invigilators))
 
-        logger.info(f"Total of tasks requested:\t {len(tasks)}")
-        logger.info(f"Number of valid results:\t {len(valid_results)}")
         return self.answers, valid_results
-
-    conduct_interview = sync_wrapper(async_conduct_interview)
 
     def _extract_valid_results(
         self, tasks, invigialtors
     ) -> Generator["Answers", None, None]:
         """Extracts the valid results from the list of results."""
 
+        # we only need to print the warning once if a task failed.
+        warning_printed = False
         warning_header = textwrap.dedent(
             """\
             WARNING: At least one question in the survey was not answered.
             """
         )
         # there should be one one invigilator for each task
-        assert len(tasks) == len(invigialtors)
-        warning_printed = False
+        assert len(self.tasks) == len(self.invigilators)
 
-        for task, invigilator in zip(tasks, invigialtors):
+        for task, invigilator in zip(self.tasks, self.invigilators):
+            logger.info(f"Iterating through task: {task}")
             if task.done():
-                if (exception := task.exception()) is not None:
+                try:
+                    result = task.result()
+                except asyncio.CancelledError:
+                    logger.info(f"Task `{task.edsl_name}` was cancelled.")
+                    result = invigilator.get_failed_task_result()
+                except Exception as exception:
                     if not warning_printed:
                         warning_printed = True
                         print(warning_header)
-                    print(
-                        f"""Task `{task.edsl_name}` failed with `{exception.__class__.__name__}`:`{exception}`."""
-                    )
+
+                    error_message = f"Task `{task.edsl_name}` failed with `{exception.__class__.__name__}`:`{exception}`."
+                    logger.error(error_message)
+                    print(error_message)
+                    traceback.print_exc()
                     # if task failed, we use the invigilator to get the failed task result
                     result = invigilator.get_failed_task_result()
                 else:
-                    result = task.result()
+                    # No exception means the task completed successfully
+                    pass
+
                 yield result
 
-    def _build_question_tasks(self, debug) -> List[asyncio.Task]:
+    def _build_question_tasks(self, debug, model_buckets) -> List[asyncio.Task]:
         """Creates a task for each question, with dependencies on the questions that must be answered before this one can be answered."""
         logger.info("Creating tasks for each question")
         tasks = []
         invigilators = []
         for question in self.survey.questions:
+            # finds dependency tasks for that question
             tasks_that_must_be_completed_before = (
                 self._get_tasks_that_must_be_completed_before(tasks, question)
             )
+            # creates the task for that question
             question_task = self._create_question_task(
-                question, tasks_that_must_be_completed_before, debug
+                question=question,
+                tasks_that_must_be_completed_before=tasks_that_must_be_completed_before,
+                model_buckets=model_buckets,
+                debug=debug,
             )
+            # adds the task to the list of tasks
             tasks.append(question_task)
             invigilators.append(self.get_invigilator(question, debug))
-        return tasks, invigilators
+        return TasksList(tasks), invigilators
 
     def _get_tasks_that_must_be_completed_before(
         self, tasks, question
@@ -221,7 +198,7 @@ class Interview:
         """Returns the tasks that must be completed before the given question can be answered.
         If a question has no dependencies, this will be an empty list, [].
         """
-        parents_of_focal_question = self.dag.get(question.question_name, [])
+        parents_of_focal_question: List[str] = self.dag.get(question.question_name, [])
         return [
             tasks[self.to_index[parent_question_name]]
             for parent_question_name in parents_of_focal_question
@@ -231,16 +208,19 @@ class Interview:
         self,
         question: Question,
         tasks_that_must_be_completed_before: List[asyncio.Task],
+        model_buckets: ModelBuckets,
         debug,
     ):
-        """Creates a task that depends on the passed-in dependencies that are awaited before the task is run.
-        The key awaitable is the `run_task` function, which is a wrapper around the `answer_question_and_record_task` method.
-        """
-        # creates the *function* that creates the task
-        create_task = QuestionTaskCreator(func=self._answer_question_and_record_task)
-        for dependency in tasks_that_must_be_completed_before:
-            create_task.append(dependency)
-        return create_task(question, debug)
+        """Creates a task that depends on the passed-in dependencies that are awaited before the task is run."""
+        task_creator = QuestionTaskCreator(
+            question=question,
+            answer_question_func=self._answer_question_and_record_task,
+            token_estimator=self._get_estimated_request_tokens,
+            model_buckets=model_buckets,
+        )
+        [task_creator.add_dependency(x) for x in tasks_that_must_be_completed_before]
+        self.task_creators[question.question_name] = task_creator
+        return task_creator.generate_task(debug)
 
     def async_timeout_handler(timeout):
         def decorator(func):
@@ -256,7 +236,7 @@ class Interview:
 
         return decorator
 
-    def get_invigilator(self, question, debug):
+    def get_invigilator(self, question, debug) -> "Invigilator":
         invigilator = self.agent.create_invigilator(
             question=question,
             scenario=self.scenario,
@@ -266,6 +246,20 @@ class Interview:
             current_answers=self.answers,
         )
         return invigilator
+
+    def _get_estimated_request_tokens(self, question) -> float:
+        """Estimates the number of tokens that will be required to run the focal task."""
+        invigilator = self.get_invigilator(question, debug=False)
+        # TODO: There should be a way to get a more accurate estimate.
+        combined_text = ""
+        for prompt in invigilator.get_prompts().values():
+            if hasattr(prompt, "text"):
+                combined_text += prompt.text
+            elif isinstance(prompt, str):
+                combined_text += prompt
+            else:
+                raise ValueError(f"Prompt is of type {type(prompt)}")
+        return len(combined_text) / 4.0
 
     @async_timeout_handler(TIMEOUT)
     async def _answer_question_and_record_task(
@@ -278,27 +272,42 @@ class Interview:
         """
         invigilator = self.get_invigilator(question, debug=debug)
         response: AgentResponseDict = await invigilator.async_answer_question()
+        # TODO: Move this back into actual agent response dict and enforce it.
         response["question_name"] = question.question_name
 
         self.answers.add_answer(response, question)
 
         self._cancel_skipped_questions(question)
 
+        # TODO: This should be forced to be a data-exchange model to cement attributes.
         return response
-    
 
-    def _cancel_skipped_questions(self, current_question):
+    def _cancel_skipped_questions(self, current_question) -> None:
         """Cancels the tasks for questions that are skipped."""
+        logger.info(f"Current question is {current_question.question_name}")
         current_question_index = self.to_index[current_question.question_name]
-        next_question = self.survey.rule_collection.next_question(q_now=current_question_index, answers=self.answers)
+        next_question = self.survey.rule_collection.next_question(
+            q_now=current_question_index, answers=self.answers
+        )
         next_question_index = next_question.next_q
-        if next_question_index > current_question_index + 1:
-            # print("We've got tasks to cancel!")
-            for i in range(current_question_index + 1, next_question_index):
+
+        def cancel_between(start, end):
+            for i in range(start, end):
+                logger.info(
+                    f"Cancelling task for question {i}; {self.tasks[i].edsl_name}"
+                )
                 self.tasks[i].cancel()
                 skipped_question_name = self.survey.question_names[i]
-                print(f"{skipped_question_name} skipped.")
-       
+                logger.info(f"{skipped_question_name} skipped.")
+
+        if next_question_index == EndOfSurvey:
+            cancel_between(current_question_index + 1, len(self.survey.questions))
+            return
+
+        if next_question_index > (current_question_index + 1):
+            cancel_between(current_question_index + 1, next_question_index)
+
+        self.tasks.status()
 
     #######################
     # Dunder methods