+++ edsl/language_models/LanguageModel.py
@@ -88,8 +88,6 @@ class RegisterLanguageModelsMeta(ABCMeta):
                 required_parameters=[("raw_response", dict[str, Any])],
                 must_be_async=False,
             )
-            # breakpoint()
-            # RegisterLanguageModelsMeta._registry[name] = cls
             RegisterLanguageModelsMeta._registry[model_name] = cls
 
     @classmethod
@@ -230,6 +228,11 @@ class LanguageModel(
 
     _model_ = None
 
+    __rate_limits = None
+    # TODO: Use the OpenAI Teir 1 rate limits
+    __default_rate_limits = {"rpm": 10_000, "tpm": 2_000_000}
+    _safety_factor = 0.8
+
     def __init__(self, crud: CRUDOperations = CRUD, **kwargs):
         """
         Attributes:
@@ -249,10 +252,35 @@ class LanguageModel(
             if key not in parameters:
                 setattr(self, key, value)
 
-        # for key, value in kwargs.items():
-        # setattr(self, key, value)
+        # TODO: This can very likely be removed
         self.api_queue = Queue()
         self.crud = crud
+
+    def __hash__(self):
+        "Allows the model to be used as a key in a dictionary"
+        return hash(self.model + str(self.parameters))
+
+    def __eq__(self, other):
+        return self.model == other.model and self.parameters == other.parameters
+
+    def _set_rate_limits(self) -> None:
+        if self.__rate_limits is None:
+            if hasattr(self, "get_rate_limits"):
+                self.__rate_limits = self.get_rate_limits()
+            else:
+                self.__rate_limits = self.__default_rate_limits
+
+    @property
+    def RPM(self):
+        "Model's requests-per-minute limit"
+        self._set_rate_limits()
+        return self._safety_factor * self.__rate_limits["rpm"]
+
+    @property
+    def TPM(self):
+        "Model's tokens-per-minute limit"
+        self._set_rate_limits()
+        return self._safety_factor * self.__rate_limits["tpm"]
 
     @staticmethod
     def _overide_default_parameters(passed_parameter_dict, default_parameter_dict):
@@ -385,11 +413,14 @@ class LanguageModel(
         try:
             dict_response = json.loads(response)
         except json.JSONDecodeError as e:
-            print("Could not load JSON. Trying to repair.")
-            print(response)
+            # TODO: Turn into logs to generate issues
             dict_response, success = await repair(response, str(e))
             if not success:
                 raise Exception("Even the repair failed.")
+
+        dict_response["cached_response"] = raw_response["cached_response"]
+        dict_response["usage"] = raw_response.get("usage", {})
+        dict_response["raw_model_response"] = raw_response
         return dict_response
 
     get_response = sync_wrapper(async_get_response)