+++ edsl/results/ResultsGGMixin.py
@@ -4,8 +4,8 @@ import tempfile
 import matplotlib.pyplot as plt
 import matplotlib.image as mpimg
 
-class ResultsGGMixin:
 
+class ResultsGGMixin:
     def ggplot2(self, ggplot_code: str, filename=None, shape="wide"):
         # Fetching DataFrame based on shape
         if shape == "long":
@@ -19,7 +19,7 @@ class ResultsGGMixin:
         # Embed the CSV data within the R script
         csv_data_escaped = csv_data.replace("\n", "\\n").replace("'", "\\'")
         read_csv_code = f"self <- read.csv(text = '{csv_data_escaped}', sep = ',')\n"
-        
+
         # Load ggplot2 library
         load_ggplot2 = "library(ggplot2)\n"
 
@@ -33,7 +33,13 @@ class ResultsGGMixin:
         # Add command to save the plot to a file
         full_r_code += f'\nggsave("{filename}", plot = last_plot(), width = 6, height = 4, device = "png")'
 
-        result = subprocess.run(['Rscript', '-'], input=full_r_code, text=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        result = subprocess.run(
+            ["Rscript", "-"],
+            input=full_r_code,
+            text=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
 
         if result.stderr:
             print("Error in R script:", result.stderr)
@@ -43,6 +49,5 @@ class ResultsGGMixin:
     def _display_plot(self, filename):
         img = mpimg.imread(filename)
         plt.imshow(img)
-        plt.axis('off')
+        plt.axis("off")
         plt.show()
-